fix: wrong `useWindowDimensions` on iPad
