+++ src/hooks/useWindowDimensions/index.ts
@@ -1,6 +1,8 @@
 import { useEffect, useState } from "react";
 import { Dimensions } from "react-native";
 
+import { WindowDimensionsEvents } from "../../bindings";
+
 import type { WindowDimensionsEventData } from "../../types";
 
 const screen = Dimensions.get("screen");
@@ -10,34 +12,20 @@ let initialDimensions: WindowDimensionsEventData = {
   height: screen.height,
 };
 
-Dimensions.addEventListener("change", (e) => {
-  initialDimensions = {
-    width: e.screen.width,
-    height: e.screen.height,
-  };
+WindowDimensionsEvents.addListener("windowDidResize", (e) => {
+  initialDimensions = e;
 });
 
-/**
- * On iOS we need to use `screen`, because this property is derived from `mainScreen.bounds.size`
- * while `window` is based on `[RCTKeyWindowValuesProxy sharedInstance].windowSize`, which can have
- * out of date values (especially when device gets rotated).
- *
- * @returns Window dimension.
- * @example
- * ```
- * const { height, window } = useWindowDimensions();
- * ```
- */
 export const useWindowDimensions = () => {
   const [dimensions, setDimensions] = useState(initialDimensions);
 
   useEffect(() => {
-    const subscription = Dimensions.addEventListener("change", (e) => {
-      setDimensions({
-        width: e.screen.width,
-        height: e.screen.height,
-      });
-    });
+    const subscription = WindowDimensionsEvents.addListener(
+      "windowDidResize",
+      (e) => {
+        setDimensions(e);
+      },
+    );
 
     // we might have missed an update between reading a value in render and
     // `addListener` in this handler, so we set it here. If there was
